Fixed duplicate user/customer data in SSO/alfa after associating existing subscription or upgrading trial subscription from dummy address.
The client needs to pass the email address to the checkAboNo API to allow the SSO to additionally check whether the subscriber ID is associated with a different account already.

And if the checkAboNo validation was correct, the client needs to pass the validated subscriber ID to registerUserAndPurchase and registerPurchase APIs, so that the SSO does not invoke registerUser towards alfa, but essentially does the opposite and imports the data from that subscriber ID from alfa into the SSO user account.

The same procedure also applies to the updateUser API when a user attempts to associate their existing subscription to their user account.

Lastly, a user should still be able to associate their existing subscription as long as there is only a "dummy address" in alfa, which refers to a partial address for trial subscriptions whose otherwise required fields have been filled up with placeholder values.  ssoFACT announces these addresses as "alfa_dummy_address" in the UserInfo now.

- https://jira.newsfactory.de/browse/HS-115
- https://jira.newsfactory.de/browse/HS-126